### 2021-11-30 #3962 @whedon
[PRE REVIEW]: hilbertmodgroup: Reduction algorithms and framework for Hilbert Modular Groups
**Submitting author:** @fredstro (<a href="http://orcid.org/0000-0002-6960-1002">Fredrik Stromberg</a>)
**Repository:** <a href="https://github.com/fredstro/hilbertmodgroup" target ="_blank">https://github.com/fredstro/hilbertmodgroup</a>
**Version:** 1.0
