### 2021-11-30 #3962 @fredstro
Hi Kristen,

I first started working on this a long time ago with a first attempt at classes etc. for Hilbert modular groups in
https://github.com/fredstro/psage/tree/master/psage/modform/hilbert
after this I have been working on a local branch in a private git repository on bitbucket.
I only copied everything over to the public GitHub repository on Nov 16 when I felt it was ready enough to be public.

Unfortunately the original repository on bitbucket was really messy with a few different projects in different directories so I decided on not trying to copy over the git log together with the code.

Best wishes
Fredrik

From: Kristen Thyng ***@***.***>
Reply to: openjournals/joss-reviews ***@***.***>
Date: Tuesday, 30 November 2021 at 17:15
Cc: Fredrik Strömberg ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: hilbertmodgroup: Reduction algorithms and framework for Hilbert Modular Groups (Issue #3962)

@fredstro<https://github.com/fredstro> Thanks for your submission! Can you give me a little history? Did you bring this code over from another location Nov 16th? Thanks.
