### 2021-12-01 #3962 @fredstro
Potential reviewers should (for ease of process) have a current SageMath installation.
From the list I would suggest: tbirkandan
Otherwise, for instance:
Samuel Lelièvre <samuel.lelievre@gmail.com>
or maybe
John Cremona  <john.cremona@gmail.com>
