### 2021-12-09 #3962 @tbirkandan
Dear Daniel,

I am experiencing some problems with my computer. Could you give me one
more day to decide? I will try to resolve the issues.

Best regards,
Tolga Birkandan

On Wed, Dec 8, 2021 at 5:41 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @JohnCremona <https://github.com/JohnCremona> and @slel
> <https://github.com/slel> - Would one or both of you be willing to review
> this submission for JOSS? (And if not, can you recommend anyone else?)
>
> JOSS is a free, open-source, community driven and developer-friendly
> online journal (no publisher is seeking to raise revenue from the volunteer
> labor of researchers!).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged.
>
> JOSS reviews involve downloading and installing the software, and
> inspecting the repository and submitted paper for key elements. See
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Editors and reviewers post comments on the Review issue, and authors
> respond to the comments and improve their submission until acceptance (or
> withdrawal, if they feel unable to satisfy the review).
>
