### 2021-12-11 #3962 @edgarcosta
Sorry, I am not able to do this.

On Fri, Dec 10, 2021 at 5:40 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @edgarcosta <https://github.com/edgarcosta>, @videlec
> <https://github.com/videlec>, @saraedum <https://github.com/saraedum>,
> @williamstein <https://github.com/williamstein> - Would one of you be
> willing to review this submission for JOSS? (And if not, can you recommend
> anyone else?)
>
> JOSS is a free, open-source, community driven and developer-friendly
> online journal (no publisher is seeking to raise revenue from the volunteer
> labor of researchers!).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged.
>
> JOSS reviews involve downloading and installing the software, and
> inspecting the repository and submitted paper for key elements. See
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Editors and reviewers post comments on the Review issue, and authors
> respond to the comments and improve their submission until acceptance (or
> withdrawal, if they feel unable to satisfy the review).
>
