### 2021-12-12 #3962 @fredstro
I’m in no hurry at the moment so if the review process takes time that’s fine.
If you have sage 9.4 installed it should be very smooth to install and run doctests etc.
