### 2021-12-15 #3996 @whedon
[REVIEW]: hilbertmodgroup: Reduction algorithms and framework for Hilbert Modular Groups
**Submitting author:** @fredstro (<a href="http://orcid.org/0000-0002-6960-1002">Fredrik Stromberg</a>)
**Repository:** https://github.com/fredstro/hilbertmodgroup
**Branch with paper.md** (empty if default branch): paper_joss
**Version:** 0.1.0
