### 2021-12-15 #3996 @danielskatz
We aim for reviews to be completed within about 2-4 weeks (though recognize that this may be delayed due to the holidays and that @videlec may be unable to start until Feb 1). Please let me know if either of you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
