### 2021-12-15 #3996 @danielskatz
@fredstro - note that the possibly missing DOIs that whedon suggests appears to be correct. Please feel free to make this change to your .bib file, then use the command `@whedon check references from branch paper_joss` to check again, and the command `@whedon generate pdf from branch paper_joss` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
