+++ paper.bib
@@ -22,7 +22,8 @@
   number=2,
   pages="31--39",
   year=2011,
-  publisher="IEEE"
+  publisher="IEEE",
+  url="https://doi.org/10.1109/mcse.2010.118"
 }
 
 @unpublished{reductionalghilbert,
