+++ paper.bib
@@ -23,6 +23,7 @@
   pages="31--39",
   year=2011,
   publisher="IEEE",
+  doi=  "10.1109/mcse.2010.118"
   url="https://doi.org/10.1109/mcse.2010.118"
 }
 
