### 2021-12-17 #3996 @tbirkandan
I installed and run the package (all doctest and all example notebooks by following [arXiv:2111.13274 [math.NT]](https://arxiv.org/abs/2111.13274)) with no problem. I just have two comments on the references of the paper (as this is not about the software, I am not opening an issue on the repository):

1) The reference of PARI/GP should be corrected according to https://pari.math.u-bordeaux.fr/faq.html (Please see the answer to the question "Need I formally cite PARI/GP in my paper?")
2) The arXiv number arXiv:2111.13274 [math.NT] can be added to the reference Strömberg, F. (2021). A reduction algorithm for Hilbert modular groups. However, as it is a preprint, I leave the decision to the author.
