### 2021-12-17 #3996 @fredstro
@tbirkandan: about the PARI reference, except the text "available from" the bibtex entry is the same as in the faq.html but the reference style used in JOSS does not output it in the right way (in particular the 'note' field is ignored).  I assume that you want me to modify the BibTex entry so that the output is in line with the example in the FAQ, i.e:
The PARI~Group, PARI/GP version \texttt{2.13.1}, Univ. Bordeaux, 2021,    \url{http://pari.math.u-bordeaux.fr/}.

@danielskatz: it seems that the note field is ignored, should I put a link in a url field to the ArXiv article?
