+++ paper.bib
@@ -1,17 +1,16 @@
 @manual{sage,
-	author = "The Sage Developers",
-	organization = "The Sage~Development Team",
+	author = "{The Sage Developers}",
 	title = "{S}age {M}athematics {S}oftware ({V}ersion 9.4)",
 	year = 2021,
-	note         = "https://www.sagemath.org"
+	url         = "https://www.sagemath.org"
 }
 
 @manual{PARI2,
-      organization = "The PARI~Group",
+      author = "{The PARI~Group}",
       title        = "PARI/GP version \texttt{2.13.1}",
       year         = 2021,
-      address      = "Univ. Bordeaux",
-      note         = "https://pari.math.u-bordeaux.fr"
+      organisation      = "Univ. Bordeaux",
+      url         = "https://pari.math.u-bordeaux.fr"
 }
 
 @article{behnel2011cython,
@@ -27,13 +26,14 @@
   url="https://doi.org/10.1109/mcse.2010.118"
 }
 
-@unpublished{reductionalghilbert,
-	author = "Strömberg, Fredrik",
-	note = "Under review by the Journal of Number Theory; Arxiv:2111.13274",
-	title = "A reduction algorithm for {H}ilbert modular groups",
-	year = 2021
+@misc{reductionalghilbert,
+      title={A Reduction Algorithm for Hilbert Modular Groups},
+      author={Strömberg, Fredrik},
+      year={2021},
+      eprint={2111.13274},
+      archivePrefix={arXiv},
+      primaryClass={math.NT}
 }
-
 @article{MR4091535,
   author =        "Quinn, Joseph and Verjovsky, Alberto",
   journal =       "Geometriae Dedicata",
