### 2021-12-17 #3996 @danielskatz
@fredstro - note that you can do this too to check the PDF when you make changes in the .md or .bib files
