### 2021-12-17 #3996 @danielskatz
And note that the [JOSS examples](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) show some ways to get things to work
