### 2021-12-17 #3996 @danielskatz
finally, I note that for the ref with the arxiv preprint - this looks fine in terms of the arxiv part, but you could change Hilbert to {H}ilbert in the title to protect that case.
