+++ paper.bib
@@ -27,7 +27,7 @@
 }
 
 @misc{reductionalghilbert,
-      title={A Reduction Algorithm for Hilbert Modular Groups},
+      title={A Reduction Algorithm for {H}ilbert Modular Groups},
       author={Strömberg, Fredrik},
       year={2021},
       eprint={2111.13274},
