### 2021-12-17 #3996 @tbirkandan
@danielskatz and @fredstro - I checked the remaining "performance" and "references" boxes. My final suggestion is to publish the final form of the paper in JOSS.
