### 2021-12-29 fredstro/hilbertmodgroup#1 @videlec
Releases
It is standard to provide releases for any library or software. As far as this Python package is concerned, the most straightforward would be to
- tag the sources with a version number (probably `0.0.0`)
- make a source distribution (`python setup.py sdist`)
- check source distribution installability (with `pip install`) and validity (using `twine check`)
- upload the source distribution on [PyPI](https://pypi.org/)

Part of the review for openjournals/joss-reviews#3996
