### 2021-12-29 fredstro/hilbertmodgroup#2 @videlec
Modern setup configuration
Python packaging is shifting toward havin a `pyproject.toml` at the root of the package that specifies the tool to install the package (for this package this would be `setuptools`). Also it becomes more common to specify the setup options in a `setup.cfg` file rather than in the `setup.py`. This way of doing is much more complete to specify build dependencies and runtime dependencies.

Part of openjournals/joss-reviews#3996
