### 2021-12-29 fredstro/hilbertmodgroup#3 @videlec
Mention hilbertmodgroup on SageMath wiki
The SageMath wiki has a [page dedicated to external packages](https://wiki.sagemath.org/SageMathExternalPackages). Please add a section for hilbertmodgroup.

Part of openjournals/joss-reviews#3996.
