### 2021-12-29 fredstro/hilbertmodgroup#4 @videlec
github workflows to run tests
[github actions](https://docs.github.com/en/actions) provide a way to run tests using the github infrastructure. This is very convenient while working on the code.

Part of openjournals/joss-reviews#3996.
