### 2021-12-29 #3996 @videlec
I opened 4 PR regarding the basic setup configuration of the project (fredstro/hilbertmodgroup#4, fredstro/hilbertmodgroup#5, fredstro/hilbertmodgroup#6, fredstro/hilbertmodgroup#7). I am starting my reading of the source code now.
