### 2022-01-04 fredstro/hilbertmodgroup#1 @fredstro
@videlec what do you think is the best way of keeping track of version, using a version number in code/separate file or using git tags and  setuptools-scm? I looked at the various options at https://packaging.python.org/en/latest/guides/single-sourcing-package-version/ but I am not sure what is the best (or modern) way.
