### 2022-01-04 fredstro/hilbertmodgroup#1 @videlec
It is convenient for users to have the version available inside the package. For example
```
>>> import numpy
>>> numpy.__version__
'1.17.4'
```
And it is convenient for you to minimize the number of places where you hardcode this number.

What I would do is to hardcode it in both `setup.py` (or `setup.cfg`) and a file `hilbertmodgroup/version.py`
```python
version = '0.0.0'
```
Then in the `__init__.py` file it would be convenient to have
```python
from .version import version as __version__
```
(for now your `__init__` file is empty but that might change in the future)

Any commit where you modify the version number would better contain **only** this change. That would be the commit you tag with the corresponding version number in the git history.
