### 2022-01-04 fredstro/hilbertmodgroup#1 @videlec
Note that there are automated tools that can handle the whole "make a new version" process such as [rever](https://regro.github.io/rever-docs/).
