### 2022-01-04 fredstro/hilbertmodgroup#1 @videlec
I did not know setuptools-scm but it definitely looks like a reasonable solution.
