### 2022-01-06 ccb02c4
Fixed codespell and pycodestyle errors
