+++ src/hilbert_modgroup/pullback.py
@@ -43,7 +43,7 @@ class HilbertPullback(SageObject):
         r"""
         Init self.
 
-        INPUT::
+        INPUT:
 
         - ``G`` - A Hilbert modular group
 
@@ -199,7 +199,7 @@ class HilbertPullback(SageObject):
         r"""
         Compute the coordinate of y=Im(z)/N(y)^(1/n) with respect to the logarithmic unit lattice.
 
-        INPUT::
+        INPUT:
 
         - ``z`` -- element of type UpperHalfPlaneProductElement
         - ``return_error_estimate`` -- boolean (default=False) set to True to return a tuple including an estimate of the error.
@@ -264,7 +264,7 @@ class HilbertPullback(SageObject):
         Reduce the point z with respect to action of units to a point where Im(z) belongs
         to a fundamental domain for the logarithmic unit lattice.
 
-        INPUT::
+        INPUT:
 
         - ``z`` -- element of type UpperHalfPlaneProductElement
         - ``return_map`` -- boolean (default=False) set to True to return the map which does the reduction.
@@ -330,7 +330,7 @@ class HilbertPullback(SageObject):
         r"""
         Return True if z is reduced with respect to the logarithmic unit lattice, otherwise return False.
 
-        INPUT::
+        INPUT:
 
         - ``z`` -- element of the type UpperHalfPlaneProductElement_class
 
@@ -380,7 +380,7 @@ class HilbertPullback(SageObject):
         r"""
         Return the Basis matrix corresponding to an integer basis of an ideal a.
 
-        INPUT::
+        INPUT:
 
         - ``a`` -- ideal or number field element.
         - ``prec`` -- integer (default=53)
@@ -433,7 +433,7 @@ class HilbertPullback(SageObject):
         Return the Basis matrix corresponding to an integer basis of an ideal a
         in terms of the standard power basis.
 
-        INPUT::
+        INPUT:
 
         - ``a`` -- ideal or number field element.
 
@@ -482,7 +482,7 @@ class HilbertPullback(SageObject):
         r"""
         Return the coordinates of x with respect to an integral basia of a.
 
-        INPUT::
+        INPUT:
 
         - ``x`` -- element of ideal a
         - ``a`` -- ideal or number field element.
@@ -525,7 +525,7 @@ class HilbertPullback(SageObject):
         r"""
         Return the Basis matrix corresponding to an integer basis of an ideal a.
 
-        INPUT::
+        INPUT:
 
         - ``a`` -- ideal or number field element.
         - ``prec`` -- integer (default=53)
@@ -570,7 +570,7 @@ class HilbertPullback(SageObject):
         L = OKz + OK embedded in R^{2n} and given by the fixed
         integral basis of the ideal a.
 
-        INPUT::
+        INPUT:
 
         - ``z`` -- element of a product of complex planes
         - ``a`` -- ideal or number field element.
@@ -891,7 +891,7 @@ class HilbertPullback(SageObject):
         r"""
         Reduce the point z with respect to the cuspidal region in a neighbourhood of the cusp .
 
-        INPUT::
+        INPUT:
 
         - ``z`` -- point in the upper half-plane.
         - ``a`` -- ideal cusp (default=None) if None use the ring of integers.
@@ -1010,7 +1010,7 @@ class HilbertPullback(SageObject):
         r"""
         Reduce ``z`` to a point in the fundamental domain.
 
-        INPUT::
+        INPUT:
 
         - ``z`` -- point in the upper half-plane
         - ``return_map`` -- boolean (default False) set to ``True`` to return the map which performed the reduction.
@@ -1049,7 +1049,7 @@ class HilbertPullback(SageObject):
         r"""
[…]
