### 2022-01-06 cfb1e57
some PEP8 fixes
