+++ src/hilbert_modgroup/hilbert_modular_group_class.py
@@ -29,6 +29,7 @@ from hilbert_modgroup.upper_half_plane import ComplexPlaneProductElement__class
 from .hilbert_modular_group_element import HilbertModularGroupElement
 
 import logging
+
 logger = logging.getLogger(__name__)
 logger.setLevel(10)
 
@@ -54,7 +55,7 @@ def is_HilbertModularGroup(x):
         sage: is_HilbertModularGroup(H)
         True
     """
-    return isinstance(x,HilbertModularGroup_class)
+    return isinstance(x, HilbertModularGroup_class)
 
 
 def HilbertModularGroup(number_field, projective=True):
@@ -131,7 +132,7 @@ class HilbertModularGroup_class(LinearMatrixGroup_generic):
 
         """
         if not is_NumberFieldOrder(base_ring) or not base_ring.number_field().is_totally_real():
-            raise ValueError("Input (={0}) can not be used to create a Hilbert modular group. Need an order of a "+
+            raise ValueError("Input (={0}) can not be used to create a Hilbert modular group. Need an order of a " +
                              "totally real number field")
         # Instance data related to elliptic elements
         self._elliptic_elements_traces = []
@@ -147,11 +148,12 @@ class HilbertModularGroup_class(LinearMatrixGroup_generic):
         self._cusp_normalizing_maps_inverse = {}
         # At the moment we only deal with full level (1)
         self._level = base_ring.ideal(1)
-        super(HilbertModularGroup_class, self).__init__(degree=Integer(2), base_ring=base_ring, special=True, sage_name=sage_name,
+        super(HilbertModularGroup_class, self).__init__(degree=Integer(2), base_ring=base_ring, special=True,
+                                                        sage_name=sage_name,
                                                         latex_string=latex_string, category=Groups().Infinite(),
                                                         invariant_form=None)
 
-    def __call__(self,a):
+    def __call__(self, a):
         r"""
         Create an element of this Hilbert modular group if possible
 
@@ -177,7 +179,7 @@ class HilbertModularGroup_class(LinearMatrixGroup_generic):
             [          0           1]
 
         """
-        return super(HilbertModularGroup_class,self).__call__(a)
+        return super(HilbertModularGroup_class, self).__call__(a)
 
     @cached_method
     def generators(self, algorithm='standard'):
@@ -242,7 +244,7 @@ class HilbertModularGroup_class(LinearMatrixGroup_generic):
         return self([0, -1, 1, 0])
 
     @cached_method
-    def T(self,a=1):
+    def T(self, a=1):
         """
         Return the element T^a = ( 1 & a // 0 & 1 ) of self.
 
@@ -268,7 +270,7 @@ class HilbertModularGroup_class(LinearMatrixGroup_generic):
         return self([1, a, 0, 1])
 
     @cached_method
-    def L(self,a):
+    def L(self, a):
         """
         Return the element L=( 1 & 0 // a & 1 ) of self.
 
@@ -317,7 +319,7 @@ class HilbertModularGroup_class(LinearMatrixGroup_generic):
             [          0 1/2*a + 1/2]
 
         """
-        return self([u, 0, 0, u**-1])
+        return self([u, 0, 0, u ** -1])
 
     def gens(self, algorithm='standard'):
         r"""
@@ -488,9 +490,9 @@ class HilbertModularGroup_class(LinearMatrixGroup_generic):
             self._ideal_cusp_representatives = []
 
             def _find_equivalent_ideal_of_minimal_norm(c):
-                for a in self.base_ring().number_field().ideals_of_bdd_norm(c.norm()-1).items():
+                for a in self.base_ring().number_field().ideals_of_bdd_norm(c.norm() - 1).items():
                     for ideala in a[1]:
[…]
