+++ src/hilbert_modgroup/pullback.py
@@ -29,17 +29,18 @@ from hilbert_modgroup.upper_half_plane import ComplexPlaneProductElement__class,
 from hilbert_modgroup.utils import upper, lower
 
 from hilbert_modgroup.pullback_cython import lattice_elements_in_box, \
-    coordinates_to_ideal_elements,find_closest_cusp, find_candidate_cusps, distance_to_cusp
+    coordinates_to_ideal_elements, find_closest_cusp, find_candidate_cusps, distance_to_cusp
 import logging
 
 log = logging.getLogger(__name__)
+
 
 class HilbertPullback(SageObject):
     r"""
     Utility class for pullback / reduction algorithms for Hilbert modular groups.
 
     """
-    def __init__(self,G):
+    def __init__(self, G):
         r"""
         Init self.
 
@@ -55,7 +56,7 @@ class HilbertPullback(SageObject):
             Pullback class for Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?
 
         """
-        if not isinstance(G,HilbertModularGroup_class):
+        if not isinstance(G, HilbertModularGroup_class):
             raise ValueError("Need a Hilbert modular group")
         self._group = G
 
@@ -195,7 +196,7 @@ class HilbertPullback(SageObject):
         """
         return self.group().base_ring().number_field().unit_group().fundamental_units()
 
-    def Y(self,z, return_error_estimate=False):
+    def Y(self, z, return_error_estimate=False):
         r"""
         Compute the coordinate of y=Im(z)/N(y)^(1/n) with respect to the logarithmic unit lattice.
 
@@ -256,10 +257,10 @@ class HilbertPullback(SageObject):
         B = self.basis_matrix_logarithmic_unit_lattice(prec=z.base_ring().prec())
         coordinate_vector = B.solve_right(log_vector, check=False)
         if return_error_estimate:
-            return vector(coordinate_vector),(B*coordinate_vector-log_vector).norm(Infinity)
+            return vector(coordinate_vector), (B*coordinate_vector-log_vector).norm(Infinity)
         return vector(coordinate_vector)
 
-    def reduce_by_units(self,z, return_map=False):
+    def reduce_by_units(self, z, return_map=False):
         r"""
         Reduce the point z with respect to action of units to a point where Im(z) belongs
         to a fundamental domain for the logarithmic unit lattice.
@@ -320,13 +321,13 @@ class HilbertPullback(SageObject):
         units = K.unit_group().gens()[1:]  # Only include the units != -1
         # To avoid overflow it is more efficient to apply the map, e.g. compute (z*u**-k)/u**k instead of z*u**-(2k)
         floors = [-floor(y/2+1/2) for y in self.Y(z)]
-        reducing_map = prod([self.group().E(u ** y) for u,y in zip(units,floors)])
+        reducing_map = prod([self.group().E(u ** y) for u, y in zip(units, floors)])
         reduced_point = z.apply(reducing_map)
         if return_map:
             return reduced_point, reducing_map
         return reduced_point
 
-    def is_reduced_by_units(self,z):
+    def is_reduced_by_units(self, z):
         r"""
         Return True if z is reduced with respect to the logarithmic unit lattice, otherwise return False.
 
@@ -519,7 +520,6 @@ class HilbertPullback(SageObject):
         B = self.basis_matrix_ideal_on_power_basis(a=a)
         return B.inverse()*x.vector()
 
-
     @cached_method
     def basis_matrix_ideal__norm(self, a=None, prec=53, row=None):
         r"""
@@ -556,7 +556,7 @@ class HilbertPullback(SageObject):
[…]
