### 2022-01-07 688fab5
Added setup.cfg and pyproject.toml to replace part of setup.py for a more modern installation procedure Also added the version.py file to gitignore to make it managed through setuptools_scm
