+++ setup.cfg
@@ -1,0 +1,35 @@
+[metadata]
+name = hilbert_modular_group
+version = file: src/hilbertmodgroup/version.py
+author = Fredrik Stromberg
+author_email = fredrik314@gmail.com
+url = https://github.com/fredstro/hilbertmodgroup
+description = Algorithms for Hilbert modular groups
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = GPL v3+
+license_file = LICENSE
+platform = any
+keywords = "hilbert modular groups, reduction algorithms"
+classifiers =
+      Development Status :: 4 - Beta
+      Intended Audience :: Science/Research
+      License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)
+      Operating System :: OS Independent
+      Programming Language :: C
+      Programming Language :: C++
+      Programming Language :: Python
+      Programming Language :: Cython
+      Topic :: Scientific/Engineering :: Mathematics
+project_urls =
+    Bug Tracker = https://github.com/fredstro/hilbertmodgroup/issues
+    Source Code = https://github.com/fredstro/hilbertmodgroup/
+
+[options]
+use_scm_version = True
+packages = hilbert_modgroup
+package_dir=
+    hilbert_modgroup=src/hilbert_modgroup
+
+[sdist]
+formats = gztar
