### 2022-01-08 f6a7983
Add a tox.ini to support various commands for testing and linting with tox.
