+++ tox.ini
@@ -1,0 +1,63 @@
+###
+### tox.ini with parts taken from the sagemath source
+###
+[tox]
+envlist = doctest, coverage, pycodestyle, relint, codespell,
+skipsdist = true
+
+[testenv]
+setenv =
+    SAGE=sage
+    HOME={toxinidir}
+whitelist_externals =
+    {env:SAGE}
+    tox
+
+commands = tox -c {toxinidir}/tox.ini -e {envname} -- {posargs}
+
+[testenv:doctest]
+description =
+   run Sage the Sage doctester (same as "sage -t")
+commands =
+    {env:SAGE} -t -p 0 {posargs:{toxinidir}/src/}
+
+[testenv:coverage]
+description =
+    give information about doctest coverage of files
+    (same as "sage --coverage[all]")
+commands =
+    {env:SAGE} --coverage {posargs:{toxinidir}/src/}
+
+[pycodestyle]
+max-line-length = 99
+statistics = True
+count = True
+
+# Not really PEP8 length but this is good enough here.
+[testenv:pycodestyle]
+description =
+    Check against PEP8 style conventions.
+deps = pycodestyle
+whitelist_externals = pycodestyle
+commands = pycodestyle {posargs:{toxinidir}/src/}
+
+[testenv:relint]
+description =
+    check whether some forbidden patterns appear
+    (includes all patchbot pattern-exclusion plugins)
+# https://github.com/codingjoe/relint
+# The patterns are in .relint.yaml
+deps = relint
+whitelist_externals = find
+commands = find {posargs:{toxinidir}/src/} -exec relint -c {toxinidir}/.relint.yaml \{\} +
+
+[testenv:codespell]
+description =
+    check for misspelled words in source code
+# https://pypi.org/project/codespell/
+deps = codespell
+commands = codespell \
+    --skip="*.so,*.c,*.cpp,.git,.idea,.ipynb_checkpoints,.tox,dist,*.egg-info" \
+    --dictionary=- \
+    {posargs:{toxinidir}/src}
+
