+++ src/hilbert_modgroup/pullback.py
@@ -5,9 +5,11 @@ AUTHORS:
 
 - Fredrik Stromberg (2021)
 
-NOTE: I know it is often a bad idea to write utility classes but I decided to do it anyway at least for the moment.
+NOTE: I know it is often a bad idea to write utility classes but I decided to \
+        do it anyway at least for the moment.
 """
-from hilbert_modgroup.hilbert_modular_group_class import HilbertModularGroup_class
+from hilbert_modgroup.hilbert_modular_group_class import \
+    HilbertModularGroup_class
 from sage.all import ZZ
 from sage.categories.sets_cat import cartesian_product
 from sage.functions.other import floor
@@ -24,12 +26,15 @@ from sage.structure.sage_object import SageObject
 from sage.matrix.all import Matrix
 
 from hilbert_modgroup.hilbert_modular_group_class import HilbertModularGroup
-from hilbert_modgroup.upper_half_plane import ComplexPlaneProductElement__class, UpperHalfPlaneProductElement__class, UpperHalfPlaneProductElement
+from hilbert_modgroup.upper_half_plane import \
+    ComplexPlaneProductElement__class,\
+    UpperHalfPlaneProductElement__class, UpperHalfPlaneProductElement
 
 from hilbert_modgroup.utils import upper, lower
 
 from hilbert_modgroup.pullback_cython import lattice_elements_in_box, \
-    coordinates_to_ideal_elements, find_closest_cusp, find_candidate_cusps, distance_to_cusp
+    coordinates_to_ideal_elements, find_closest_cusp, find_candidate_cusps, \
+    distance_to_cusp
 import logging
 
 log = logging.getLogger(__name__)
@@ -37,7 +42,7 @@ log = logging.getLogger(__name__)
 
 class HilbertPullback(SageObject):
     r"""
-    Utility class for pullback / reduction algorithms for Hilbert modular groups.
+    Utility class for pullback/reduction algorithms for Hilbert modular groups.
 
     """
     def __init__(self, G):
@@ -50,10 +55,10 @@ class HilbertPullback(SageObject):
 
         EXAMPLES:
 
-            sage: from hilbert_modgroup.all import HilbertModularGroup, HilbertPullback
+            sage: from hilbert_modgroup.all import *
             sage: H = HilbertModularGroup(5)
             sage: HilbertPullback(H)
-            Pullback class for Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?
+            Pullback class ... polynomial x^2 - 5 with a = 2.236067977499790?
 
         """
         if not isinstance(G, HilbertModularGroup_class):
@@ -66,11 +71,12 @@ class HilbertPullback(SageObject):
 
          EXAMPLES::
 
-            sage: from hilbert_modgroup.all import HilbertModularGroup, HilbertPullback
+            sage: from hilbert_modgroup.all import HilbertModularGroup, \
+                                                    HilbertPullback
             sage: H = HilbertModularGroup(5)
             sage: P = HilbertPullback(H)
             sage: str(P)
-            'Pullback class for Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?'
+            'Pullback class ... polynomial x^2 - 5 with a = 2.236067977499790?'
 
         """
         return f"Pullback class for {self._group}"
@@ -81,11 +87,11 @@ class HilbertPullback(SageObject):
 
          EXAMPLES::
 
-            sage: from hilbert_modgroup.all import HilbertModularGroup, HilbertPullback
+            sage: from hilbert_modgroup.all import *
             sage: H = HilbertModularGroup(5)
             sage: P = HilbertPullback(H)
             sage: P
[…]
