+++ src/hilbert_modgroup/all.py
@@ -1,4 +1,4 @@
 from .hilbert_modular_group_class import HilbertModularGroup, is_HilbertModularGroup
 from .pullback import HilbertPullback
-from .upper_half_plane import UpperHalfPlaneProductElement, ComplexPlaneProductElement, ComplexPlaneProduct, \
-    UpperHalfPlaneProduct
+from .upper_half_plane import UpperHalfPlaneProductElement, ComplexPlaneProductElement, \
+    ComplexPlaneProduct, UpperHalfPlaneProduct
