+++ src/hilbert_modgroup/hilbert_modular_group_class.py
@@ -5,9 +5,9 @@ EXAMPLES::
 
     sage: from hilbert_modgroup.all import HilbertModularGroup
     sage: HilbertModularGroup(5)
-    Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?
+    Hilbert Modular Group ... x^2 - 5 with a = 2.236067977499790?
     sage: HilbertModularGroup(QuadraticField(5))
-    Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?
+    Hilbert Modular Group ... x^2 - 5 with a = 2.236067977499790?
 
 
 AUTHORS:
@@ -65,8 +65,9 @@ def HilbertModularGroup(number_field, projective=True):
 
         INPUT:
 
-        - ``number_field`` (NumberField) -- a totally real number field or positive integer. If a positive integer D is specified
-            then the number field $Q(\sqrt(D))$ is used.
+        - ``number_field`` (NumberField) -- a totally real number field or positive integer.
+                                            If a positive integer D is specified
+                                            then the number field $Q(\sqrt(D))$ is used.
         - ``projective`` (bool) - True if you want PSL(2,K) and False for SL(2,K)
 
 
@@ -74,18 +75,18 @@ def HilbertModularGroup(number_field, projective=True):
 
             sage: from hilbert_modgroup.all import HilbertModularGroup
             sage: HilbertModularGroup(5)
-            Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?
+            Hilbert Modular Group ... x^2 - 5 with a = 2.236067977499790?
             sage: HilbertModularGroup(QuadraticField(5))
-            Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?
+            Hilbert Modular Group ... x^2 - 5 with a = 2.236067977499790?
 
 
         TESTS::
 
             sage: from hilbert_modgroup.all import HilbertModularGroup
             sage: HilbertModularGroup(5)
-            Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?
+            Hilbert Modular Group ... x^2 - 5 with a = 2.236067977499790?
             sage: HilbertModularGroup(QuadraticField(5))
-            Hilbert Modular Group PSL(2) over Maximal Order in Number Field in a with defining polynomial x^2 - 5 with a = 2.236067977499790?
+            Hilbert Modular Group ... x^2 - 5 with a = 2.236067977499790?
 
 
     """
@@ -105,7 +106,8 @@ def HilbertModularGroup(number_field, projective=True):
 
 class HilbertModularGroup_class(LinearMatrixGroup_generic):
     r"""
-    Class for Hilbert modular groups, here defined as either PSL(2) (default) or  SL(2) over rings of integers in totally real number fields.
+    Class for Hilbert modular groups, here defined as either PSL(2) (default) or  SL(2)
+    over rings of integers in totally real number fields.
 
 
     """
@@ -123,17 +125,17 @@ class HilbertModularGroup_class(LinearMatrixGroup_generic):
         - ``latex_string`` - string
 
         EXAMPLES::
-            sage: from hilbert_modgroup.hilbert_modular_group_class import HilbertModularGroup_class
+            sage: from hilbert_modgroup.hilbert_modular_group_class import *
             sage: OK=QuadraticField(5).OK()
             sage: name = f'Hilbert Modular Group PSL(2) over {OK}'
             sage: ltx = f'PSL(2, {latex(OK)})'
[…]

[…]
