### 2022-01-08 7b68558
change to number_field degree instead of base_ring degree to avoid potential caching issues
