### 2022-01-17 #3996 @fredstro
@danielskatz @videlec  all good. I just got swamped with teaching/admin work... I have done everything suggested by @videlec  except the github actions which I am still working on. (and I have only uploaded it to test.pypi so far)
