### 2022-01-18 fredstro/hilbertmodgroup#2 @fredstro
I have removed everything I can from the setup.py and put into the pyproject.toml and setup.cfg.
As far as I can tell the Cython extension modules still need to be in setup.py
