### 2022-01-19 e8d973c
Create docker-image.yml
