+++ .github/workflows/docker-image.yml
@@ -1,0 +1,20 @@
+name: Docker Image CI with tox
+
+on:
+  push:
+    branches: [ main develop ]
+  pull_request:
+    branches: [ main develop ]
+
+jobs:
+
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Build the Docker image
+      run: docker build . --file Dockerfile --tag my-image-name:$(date +%s)
+    - name: Run tox
+      run: docker run -it --init hilbertmodgroup tox-all
