+++ .github/workflows/docker-image.yml
@@ -2,9 +2,13 @@ name: Docker Image CI with tox
 
 on:
   push:
-    branches: [ main develop ]
+    branches:
+    - main
+    - develop
   pull_request:
-    branches: [ main develop ]
+    branches:
+    - main
+    - develop
 
 jobs:
 
