### 2022-01-19 2c8d8e5
Added forgotten Dockerfile and entrypoint files to git
