### 2022-01-19 629449b
add current branch to the github action workflow and print the branch in entrypoint.sh
