+++ .github/workflows/docker-image.yml
@@ -21,4 +21,4 @@ jobs:
     - name: Build the Docker image
       run: docker build . --file Dockerfile --tag my-image-name:$(date +%s)
     - name: Run tox
-      run: docker run -it --init hilbertmodgroup tox-all
+      run: docker run -it --init hilbertmodgroup tox-all ${GITHUB_REF##*/}
