### 2022-01-19 b59b250
removed tty input from the docker command. Still keep output '-t" so that output gets colored.
