### 2022-01-19 f2d690d
updated docker image name
