+++ .github/workflows/docker-image.yml
@@ -19,6 +19,6 @@ jobs:
     steps:
     - uses: actions/checkout@v2
     - name: Build the Docker image
-      run: docker build . --file Dockerfile --tag my-image-name:$(date +%s)
+      run: docker build . --file Dockerfile --tag hilbertmodgroup:$(date +%s)
     - name: Run tox on current commit
       run: docker run -t --init hilbertmodgroup tox-all ${{ github.sha }}
