### 2022-01-19 21be0e6
added date to docker image at run
