### 2022-01-19 9ef8905
exclude c files from the source distribution
