### 2022-01-19 6456b42
Updated entrypoint script and makefile
