+++ entrypoint.sh
@@ -1,6 +1,7 @@
-#!/bin/sh
+#!/bin/bash
 
-branch="${2:=develop}"
+branch="${2:-develop}"
+
 git pull -ff origin $branch
 echo "Pulling branch: $branch"
 case $1 in
@@ -8,7 +9,7 @@ case $1 in
       echo "Docker container running Sage doctests for the hilbertmodgroup package."
       sage -t src
       ;;
-    tox-all)
+    tox)
       echo "Docker container running tox with 'doctest', 'coverage', 'pycodestyle-minimal', 'relint', 'codespell'"
       sage -python -m tox src
       ;;
@@ -16,7 +17,7 @@ case $1 in
       echo "Docker container with Jupyter Notebook interface to run example notebooks."
       echo "NOTE: The Jupyter Notebook server is only accessible using the URL from outside the container."
       sage -n jupyter --no-browser --ip='0.0.0.0' --port=8888\
-                            --notebook-dir=/home/sage/hilbertmodgroup/examples\
+                            --notebook-dir=examples\
                             --NotebookApp.custom_display_url=http://127.0.0.1:8888\
                             --NotebookApp.use_redirect_file=False\
                             --NotebookApp.browser=x-www-browser
