### 2022-01-19 b275366
updated tox command in github actions
