+++ .github/workflows/docker-image.yml
@@ -24,4 +24,4 @@ jobs:
     - name: Build the Docker image
       run: docker build . --file Dockerfile --tag hilbertmodgroup:${{ steps.date.outputs.date }}
     - name: Run tox on current commit
-      run: docker run -t --init hilbertmodgroup:${{ steps.date.outputs.date }} tox-all ${{ github.sha }}
+      run: docker run -t --init hilbertmodgroup:${{ steps.date.outputs.date }} tox ${{ github.sha }}
