### 2022-01-26 fredstro/hilbertmodgroup#1 @fredstro
I decided to use setuptools-scm and at the moment the following works

```
from hilbert_modgroup.version import version as __version__
```
