### 2022-01-26 fredstro/hilbertmodgroup#1 @fredstro
I have also made a source distribution etc. and uploaded the distribution to pypi: https://pypi.org/project/hilbert-modular-group/
It can be installed with `sage -pip install hilbert-modular-group`
