### 2022-01-26 fredstro/hilbertmodgroup#4 @fredstro
I have added a github action workflow to run tox in a docker container with a number of tests (basically taken from the sage's tox configuration):
- doctests
- coverage
- pycodestyle
- relint
- codespell
