### 2022-01-26 fredstro/hilbertmodgroup#4 @videlec
Very nice.
