### 2022-02-14 b02e920
Upgrade setuptools_scm and get version from attr instead of file
