+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = hilbert_modular_group
-version = file: src/hilbertmodgroup/version.py
+version = attr: hilbert_modgroup.__version__
 author = Fredrik Stromberg
 author_email = fredrik314@gmail.com
 url = https://github.com/fredstro/hilbertmodgroup
