+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = [
     "setuptools>=49.6.0",
     "wheel>=0.36.2",
     "cython>=0.29.21, <1.0",
-    "setuptools_scm[toml]>=3.4"
-    ]
+    "setuptools_scm[toml]>=6.2",
+        ]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
