### 2022-02-14 9f4e243
enable choice of branch to build for docker images
