+++ Dockerfile
@@ -3,7 +3,8 @@ RUN sudo apt-get update
 RUN sudo apt-get -y install git
 RUN git clone https://github.com/fredstro/hilbertmodgroup.git
 WORKDIR "hilbertmodgroup"
-RUN git config pull.rebase false && git checkout -b develop
+ARG GIT_BRANCH=develop
+RUN git config pull.rebase false && git checkout $GIT_BRANCH
 RUN sage setup.py install
 COPY entrypoint.sh /entrypoint.sh
 ENTRYPOINT ["/entrypoint.sh"]
