### 2022-02-21 908b553
Hide compilation warnings that are present in code outside this project to make it easier to spot actual errors in this code.
