### 2022-02-21 1b1f028
update manifest to explicitly include .pyx files since the version of setuptools in sage 9.5 otherwise excludes them
