+++ MANIFEST.in
@@ -1,2 +1,9 @@
 exclude src/hilbert_modgroup/hilbert_modular_group_element.c
 exclude src/hilbert_modgroup/upper_half_plane.c
+exclude src/hilbert_modgroup/pullback_cython.cpp
+include src/hilbert_modgroup/hilbert_modular_group_element.pyx
+include src/hilbert_modgroup/pullback_cython.pyx
+include src/hilbert_modgroup/upper_half_plane.pyx
+include src/hilbert_modgroup/upper_half_plane.pxd
+
+
