### 2022-02-21 5ddd3e8
Restructure the makefile and dockerfile to allow for setting command line arguments by name in the make command instead of relying on arguments being in order. also added option to run the local source in docker instead of always pulling from gitHub
