### 2022-02-21 63a1255
updated readme to include examples of make commands
