+++ README.md
@@ -9,8 +9,8 @@ and is dependent on SageMath.
 
 ## Installation
 ### Using sage pip
-This package needs to be installed in the virtual environment provided by SageMath and it is therefore necessary 
-to run the install command 
+This package needs to be installed in the virtual environment provided by SageMath, and it is therefore necessary 
+to run the following command 
 ```console
 $ sage -pip install hilbert-modular-group
 ```
@@ -24,10 +24,15 @@ $ cd hilbertmodgrup
 $ make install
 ```
 
+### Docker
+If you do not have SageMath installed, but you have docker you can use install this package
+in a docker container built and executed using e.g. `make docker-sage` or `make docker-examples`
+
+
 ## Usage
 The package can be imported and used as any other package. 
 For example, to find the reduction of the point given by [1+i,1+i] in H^2 
-with respect to the Hilbert modular group of Q adjoint square-root of 5 write: 
+with respect to the Hilbert modular group of Q joint by square-root of 5 write: 
 
 ```
 sage: from hilbert_modgroup.all import *
@@ -49,7 +54,16 @@ the `/examples` directory which contains Jupyter notebook with more extensive
 examples corresponding to the paper
 "Reduction Algorithms for Hilbert Modular Groups" by F. Stromberg. (Reference to appear)
 
-## Additional Commands
+## Examples
+
+The directory `/examples` contains Jupyter notebooks with example code to illustrate the interface and functionality of this package. 
+You can either open them manually from SageMath or run one of the following commands:
+`make examples`
+`make docker-examples`
+which will start up a Jupyter notebook server from sagemath either locally or in a docker container. 
+
+## Development and testing
+
 The make file `Makefile` contains a number of useful commands that you can run using 
 ```console
 $ make <command>
@@ -61,26 +75,46 @@ The following commands are run in your local SagMath environment:
 4. `clean` -- remove all build and temporary files
 5. `test` -- run sage's doctests (same as `sage -t src/*`)
 6. `examples` -- run a Jupyter notebook with the SageMath kernel initialised at the `/examples` directory.
-7. `tox` -- run tox with all environments: `doctest`, `coverage`, `pycodestyle-minimal`, `relint`, `codespell`
+7. `tox` -- run `sage -tox` with all environments: `doctest`, `coverage`, `pycodestyle-minimal`, `relint`, `codespell`
+   Note: If your local SageMath installation does not contain tox this will run `sage -pip install tox`.
 
 The following commands are run in an isolated docker container 
 and requires docker to be installed and running:
-1. `docker` -- build a docker container with the tag `hilbertmodgroup`
+1. `docker` -- build a docker container with the tag `hilbertmodgroup-{GIT_BRANCH}`
 2. `docker-rebuild` -- rebuild the docker container without cache
 3. `docker-test` -- run SageMath's doctests in the docker container
 4. `docker-examples` -- run a Jupyter notebook with the SageMath kernel initialised at the `/examples` directory 
-  and exposing the notebook at http://127.0.0.1:8888
-5. `docker-tox` -- run tox with all environments: `doctest`, `coverage`, `pycodestyle-minimal`, `relint`, `codespell`
+  and exposing the notebook at http://127.0.0.1:8888. The port used can be modified by 
+5. `docker-tox` -- run tox with all environments: `doctest`, `coverage`, `pycodestyle-minimal`, `relint`, `codespell`. 
 6. `docker-shell` -- run a shell in a docker container
 7. `docker-sage` -- run a sage interactive shell in a docker container
 
+The following command-line parameters are available 
[…]
