+++ CHANGELOG.md
@@ -1,0 +1,15 @@
+# Changelog
+
+
+## v.0.0.3
+- Added Changelog (this file)
+- Improved handling of arguments in make and Docker.
+- Updated README.md with examples
+- 
+## v.0.0.2
+- Updated docker commands
+- Added gitHub actions 
+- Added tox environments
+
+## v.0.0.0
+- Initial release
