+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Changelog
 
 
-## v.0.0.3
+## v.0.0.2
 - Added Changelog (this file)
 - Improved handling of arguments in make and Docker.
 - Updated README.md with examples
 - 
-## v.0.0.2
+## v.0.0.1
 - Updated docker commands
 - Added gitHub actions 
 - Added tox environments
