### 2022-02-22 fredstro/hilbertmodgroup#5 @videlec
fragile SAGE_LIB in setup.py
The assumption about the localtion of `SAGE_LIB` might be wrong
```
SAGE_LIB = f"{SAGE_LOCAL}/lib/python{PYTHON_VERSION}/site-packages"
```
On my system the compiled version of sage does not satisfy the above
```
$ sage-dev -python -c 'from sage.env import SAGE_LOCAL, SAGE_LIB; print(f"{SAGE_LOCAL}\n{SAGE_LIB}")'
/opt/sage/local
/opt/sage/local/var/lib/sage/venv-python3.10/lib/python3.10/site-packages
```
