### 2022-02-22 fredstro/hilbertmodgroup#6 @videlec
Parent.__call__
At https://github.com/fredstro/hilbertmodgroup/blob/main/src/hilbert_modgroup/upper_half_plane.pyx#L236 and https://github.com/fredstro/hilbertmodgroup/blob/main/src/hilbert_modgroup/hilbert_modular_group_class.py#L160 you overwrite the default `__call__` of the sage `Parent` class. Is there any reason for that? Doing so would prevent any coercion to happen.

Part of https://github.com/openjournals/joss-reviews/issues/3996
