### 2022-02-22 fredstro/hilbertmodgroup#7 @videlec
pyflakes
In `hilbert_modular_group_element.pyx`
```
hilbert_modular_group_element.pyx:18:1 'sage.matrix.matrix_generic_dense.Matrix_generic_dense' imported but unused
hilbert_modular_group_element.pyx:21:1 'hilbert_modgroup.upper_half_plane.ComplexPlaneProductElement' imported but unused
hilbert_modular_group_element.pyx:22:1 'hilbert_modgroup.upper_half_plane.UpperHalfPlaneProductElement__class' imported but unused
```
and in `upper_half_plane.pyx`
```
upper_half_plane.pyx:13:1 'sage.structure.element.MultiplicativeGroupElement' imported but unused
upper_half_plane.pyx:14:1 'sage.structure.richcmp.richcmp' imported but unused
upper_half_plane.pyx:15:1 'sage.rings.all.ZZ' imported but unused
upper_half_plane.pyx:16:1 'sage.rings.infinity.infinity' imported but unused
upper_half_plane.pyx:19:1 'sage.matrix.matrix_space.MatrixSpace' imported but unused
upper_half_plane.pyx:20:1 'sage.matrix.matrix_generic_dense.Matrix_generic_dense' imported but unused
upper_half_plane.pyx:21:1 'sage.misc.cachefunc.cached_method' imported but unused
upper_half_plane.pyx:22:1 'sage.structure.unique_representation.UniqueRepresentation' imported but unused
upper_half_plane.pyx:28:1 'sage.categories.semigroups.Semigroups' imported but unused
upper_half_plane.pyx:29:1 'cysignals.memory.sig_free' imported but unused
upper_half_plane.pyx:29:1 'cysignals.memory.sig_malloc' imported but unused
upper_half_plane.pyx:29:1 'cysignals.memory.check_allocarray' imported but unused
upper_half_plane.pyx:30:1 redefinition of unused 'SageObject' from line 17
upper_half_plane.pyx:33:1 redefinition of unused 'vector' from line 11
upper_half_plane.pyx:880:9 redefinition of unused 'infinity' from line 16
upper_half_plane.pyx:1608:13 local variable 'maxd' is assigned to but never used
upper_half_plane.pyx:1610:30 f-string is missing placeholders
```

Part of https://github.com/openjournals/joss-reviews/issues/3996
