### 2022-02-22 fredstro/hilbertmodgroup#8 @videlec
TestSuite
sage has some automated tests to check that categories, inheritance, etc are correctly implemented. You can run it via
```
sage: from hilbert_modgroup.upper_half_plane import ComplexPlaneProduct
sage: TestSuite(ComplexPlaneProduct(2)).run()
```
The above fails because
- category is not initialized
- pickling fails

Part of https://github.com/openjournals/joss-reviews/issues/3996
