### 2022-02-22 #3996 @videlec
Sorry for the delay. I opened a new issue about the setup configuration (https://github.com/fredstro/hilbertmodgroup/issues/5) which prevented me from installing hilbertmodgroup in my compiled version of sage. I started looking at the code and opened 3 issues (https://github.com/fredstro/hilbertmodgroup/issues/6, https://github.com/fredstro/hilbertmodgroup/issues/7 and https://github.com/fredstro/hilbertmodgroup/issues/8).
