### 2022-02-22 fredstro/hilbertmodgroup#5 @fredstro
@videlec  do you have any suggestion for fixing this?

Importing SAGE_LIB from sage.env is fine for just doing `sage setup.py install` but for a source installation
i.e. `sage -python setup.py sdist` followed by `sage -pip install sdist/*` it doesn't work due to build-isolation .
The only way around seems to use the --no-build-isolation flag, i.e.
`sage -pip install dist/* --no-build-isolation`
but I don't think it is reasonable to ask people to use a special flag, or what do you think?
