### 2022-02-22 fredstro/hilbertmodgroup#5 @videlec
By providing the `SAGE_LIB` you are anyway bypassing silently the build isoliation. The only way to support the build isolation would be to set SageMath as a build dependency which is of course not viable.

The `--no-build-isolation` looks like a better option. Isn't it possible to force this option inside `setup.py` and or `setup.cfg`?

You could possibly ask the question on [sage-devel](https://groups.google.com/g/sage-devel) as I am not an expert of these Python packaging subtleties.
