### 2022-02-22 fredstro/hilbertmodgroup#5 @videlec
The discussion at https://github.com/pypa/pip/issues/5033 is relevant.
