### 2022-02-22 fredstro/hilbertmodgroup#5 @videlec
Also https://github.com/pypa/pip/issues/6718
