### 2022-03-17 532fe91
Fix coercion and use of arithmetic from parent classes in ComplexPlaneProduct__class and UpperHalfPlaneProduct__class and their element classes.
