+++ src/hilbert_modgroup/upper_half_plane.pyx
@@ -20,7 +20,9 @@ from sage.matrix.matrix_space import MatrixSpace
 from sage.matrix.matrix_generic_dense cimport Matrix_generic_dense
 from sage.misc.cachefunc import cached_method
 from sage.structure.unique_representation import UniqueRepresentation
+from sage.structure.element cimport Element
 from sage.structure.parent import Parent
+from sage.structure.element cimport parent
 
 from sage.rings.complex_mpfr cimport ComplexNumber
 from sage.rings.complex_mpc cimport MPComplexNumber, MPComplexField_class
@@ -95,28 +97,35 @@ def UpperHalfPlaneProductElement(z, **kwds):
         sage: from hilbert_modgroup.all import UpperHalfPlaneProductElement
         sage: UpperHalfPlaneProductElement([1+I,1+I])
         [1.00000000000000 + 1.00000000000000*I, 1.00000000000000 + 1.00000000000000*I]
-        sage: UpperHalfPlaneProductElement([1+I,1+I,1+I,1+I])
-        [1.00000000000000 + 1.00000000000000*I, 1.00000000000000 + 1.00000000000000*I, 1.00000000000000 + 1.00000000000000*I, 1.00000000000000 + 1.00000000000000*I]
-
-    TESTS:
-
-        sage: UpperHalfPlaneProductElement([1,1])
+        sage: set(UpperHalfPlaneProductElement([1+I]*10))
+        {1.00000000000000 + 1.00000000000000*I}
+        sage: len(UpperHalfPlaneProductElement([1+I]*10))
+        10
+        sage: UpperHalfPlaneProductElement([1,1-I])
         Traceback (most recent call last):
         ...
-        ValueError: Point [1.00000000000000, 1.00000000000000] not in upper half-plane!
+        ValueError: Point [1.00000000000000, 1.00000000000000 - 1.00000000000000*I] not in upper half-plane!
 
     """
     if isinstance(z,UpperHalfPlaneProductElement__class):
+        parent = kwds.get('parent')
+        if parent is z.parent():
+            return z
+        if parent:
+            return UpperHalfPlaneProductElement__class(list(z), parent=parent)
         return z
     prec = kwds.get('prec',getattr(z,'prec',lambda : 53)())
+    if hasattr(z,'value'):
+        z = z.value()
     if is_NumberFieldElement(z):
         z = z.complex_embeddings(prec)
-    if isinstance(z,list) and not isinstance(z[0],(ComplexNumber,MPComplexNumber)):
+    if isinstance(z,list) and not isinstance(z[0], (ComplexNumber, MPComplexNumber)):
        z  = [MPComplexField(prec)(x) for x in z]
     elif not isinstance(z,list) and kwds.get('degree',0)>0:
         z = [MPComplexField(prec)(z)]*kwds.get('degree')
-
-    return UpperHalfPlaneProductElement__class(z)
+    if 'parent' not in kwds:
+        kwds['parent'] = UpperHalfPlaneProduct(degree=len(z))
+    return UpperHalfPlaneProductElement__class(z,**kwds)
 
 def ComplexPlaneProductElement(z,**kwds):
     """
@@ -147,6 +156,11 @@ def ComplexPlaneProductElement(z,**kwds):
 
     """
     if isinstance(z,ComplexPlaneProductElement__class):
+        parent = kwds.get('parent')
+        if parent is z.parent():
+            return z
+        if parent:
+            return ComplexPlaneProductElement__class(list(z), parent=parent)
         return z
     # Get precision in the first hand from kwds, second from z and third set default to 53 bits
     prec = kwds.get('prec',getattr(z,'prec',lambda : 53)())
[…]
