### 2022-03-17 c2c9477
Add support for pickling and add TestSuite.run to all classes.
