### 2022-03-17 7f89a3b
Change the ProductOfUpperHalfPlane Class to allow for points on the boundary, i.e. with imaginary part 0. This makes sense from a practical standpoint even if theoretically it should probably be a separate subclass of ComplexPlaneProduct__class
