### 2022-03-17 2d2f4d7
Make sure that application of a matrix in Hilbert modular group preserves the upper half-plane class
