+++ src/hilbert_modgroup/upper_half_plane.pyx
@@ -672,6 +672,10 @@ cdef class ComplexPlaneProductElement__class(Element):
             [1.00000000000000 + 1.00000000000000*I, 2.00000000000000 + 3.00000000000000*I]
             sage: z.is_in_upper_half_plane()
             True
+            sage: z=ComplexPlaneProductElement([CC(1,0),CC(2,3)]); z
+            [1.00000000000000, 2.00000000000000 + 3.00000000000000*I]
+            sage: z.is_in_upper_half_plane()
+            True
             sage: z=ComplexPlaneProductElement([CC(1,1),CC(2,-1)]); z
             [1.00000000000000 + 1.00000000000000*I, 2.00000000000000 - 1.00000000000000*I]
             sage: z.is_in_upper_half_plane()
@@ -679,6 +683,28 @@ cdef class ComplexPlaneProductElement__class(Element):
 
         """
         return bool(self._is_in_upper_half_plane)
+
+    def as_upper_half_plane_element(self):
+        r"""
+        Return a copy of self with type UpperHalfPlaneProductElement__class
+
+        EXAMPLES::
+
+            sage: from hilbert_modgroup.all import ComplexPlaneProductElement
+            sage: z=ComplexPlaneProductElement([CC(1,1),CC(2,3)]); z
+            [1.00000000000000 + 1.00000000000000*I, 2.00000000000000 + 3.00000000000000*I]
+            sage: z.as_upper_half_plane_element()
+            [1.00000000000000 + 1.00000000000000*I, 2.00000000000000 + 3.00000000000000*I]
+            sage: z=ComplexPlaneProductElement([CC(1,1),CC(2,-1)]); z
+            [1.00000000000000 + 1.00000000000000*I, 2.00000000000000 - 1.00000000000000*I]
+            sage: z.as_upper_half_plane_element()
+            Traceback (most recent call last):
+            ...
+            ValueError: Can not convert self to element in product of upper half-planes.
+        """
+        if not self.is_in_upper_half_plane():
+            raise ValueError("Can not convert self to element in product of upper half-planes.")
+        return UpperHalfPlaneProductElement(self._z)
 
     def is_zero(self):
         r"""
@@ -831,33 +857,10 @@ cdef class ComplexPlaneProductElement__class(Element):
             sage: z[1]
             2.00000000000000 - 1.00000000000000*I
         """
-        if isinstance(i,(int,Integer)) and i >= 0 and i < self._degree:
+        if isinstance(i, (int, Integer)) and 0 <= i < self._degree:
             return self._z[i]
         else:
             raise IndexError
-
-    # def __reduce__(self):
-    #     return (Hn,(self._xlist,self._ylist,self._degree,self._prec,self._verbose))
-
-    # cdef c_new(self,list x,list y):
-    #     self._x = NULL; self._y=NULL
-    #     self._x = <double*>sig_malloc(sizeof(double)*self._degree)
-    #     if self._x==NULL:
-    #         raise MemoryError
-    #     self._y = <double*>sig_malloc(sizeof(double)*self._degree)
-    #     if self._y==NULL:
-    #         raise MemoryError
-    #     cdef int i
-    #     for i in range(self._degree):
-    #         self._x[i] = <double>x[i]
-    #         self._y[i] = <double>y[i]
-    #         if self._verbose>1:
-    #             print "x[{0}]={1}".format(i,x[i])
-    #             print "y[{0}]={1}".format(i,y[i])
-    #         if y[i]<0:
-    #             raise ValueError,"Not in H^n^*! y[{0}]={1}".format(i,y[i])
-    #     if self._verbose>0:
-    #         print "allocated x and y!"
 
[…]
