### 2022-03-17 ae3b824
small fixes suggested by pyflakes and an instance of explicit coercion
