### 2022-03-17 cc99471
add flake8 to tox
