+++ tox.ini
@@ -2,7 +2,7 @@
 ### tox.ini with parts taken from the sagemath source
 ###
 [tox]
-envlist = doctest, coverage, pycodestyle, relint, codespell,
+envlist = doctest, coverage, pycodestyle, relint, codespell, flake8
 skipsdist = true
 
 [testenv]
@@ -61,3 +61,17 @@ commands = codespell \
     --dictionary=- \
     {posargs:{toxinidir}/src}
 
+[flake8]
+max-line-length = 100
+filename = *.py, *.pyx
+max_complexity = 10
+per-file-ignores =
+    __init__.py:F401
+    all.py:F401
+    *.pyx:E999
+
+[testenv:flake8]
+description =
+    check for misspelled words in source code
+deps = flake8
+commands = flake8 {posargs:{toxinidir}/src/}
