### 2022-03-17 fredstro/hilbertmodgroup#7 @fredstro
I have fixed these and I added flake8 to the tox configuration
but I couldn't figure out how to make pyflakes (through flake8) detect these issues in the Cython files.
Somehow it only flags syntax errors from "cimport" and once I add an ignore flag to those no other errors are detected.
