### 2022-03-17 fredstro/hilbertmodgroup#8 @fredstro
I have fixed pickling (and coercion) and added a TestSuite().run() to all classes.
