### 2022-03-17 fredstro/hilbertmodgroup#6 @fredstro
I was actually intentionally trying to avoid Sage's coercion framework since I find it extremely mes for extension types (unless they happen to fit very well within an existing category). For instance, I couldn't make coercion work without adding a "construction" function which just returns None...

However, I took some time to do it and have now added explicit coercion to the ComplexPlaneProduct__class and UpperHalfPlaneProduct __class and their respective Element classes.
I am sure that someone that knows this better can improve on the logic but it works as intended now with all  builtin arithmetic functions
