### 2022-03-17 fredstro/hilbertmodgroup#5 @fredstro
I am still looking into this but it appears it is not possible to bypass the build isolation from any configuration file.

It seems to be an explicit intention in the current pip development that package authors should not be able to do this, only end users.
The only option seems to be to either
1) Make an spkg package instead
2) Ask users to supply a '--no-build-isolation' flag
3) Try some other method to locate the library files in the file system (look at all possible relative locations to SAGE_LOCAL)

What I have tried and unfortunately doesn't work is to use something like

cmd = "import sage; from sage.env import SAGE_LIB; print(SAGE_LIB)"
sage = shutil.which('sage')
proc = subprocess.run([sage, '-c', cmd], capture_output=True)
SAGE_LIB = proc.stdout.decode('utf8')

It works from a standard python prompty but somehow the isolated environment that pip creates makes it so that the sage library is not found even from within the sage executable...

I still haven't asked on sage-devel but will try to do that soon. I just wanted to try all options I could think of myself first...
