### 2022-03-17 fredstro/hilbertmodgroup#7 @videlec
Indeed, it is annoying that flake8 does not proposes a plugin for Cython files. But this is a problem beyond hilbertmodgroup.
