### 2022-03-21 #3996 @danielskatz
👋 @fredstro & @videlec - another ping on the questions immediately above...
