### 2022-03-21 fredstro/hilbertmodgroup#5 @fredstro
@videlec What do you think about Matthias Koeppes response (https://groups.google.com/g/sage-devel/c/QtfFtjdePcY)?

I am thinking that maybe I should aim at an install through:
 `sage -i hilbert-modular-group`
and have the "standard" pip install requiring the --no-build-isolation flag, i.e.
`sage -pip install --no-build-isolation hilbert-modular-group`
(or having the SAGE_LIBRARY environment variable set).
