### 2022-03-21 fredstro/hilbertmodgroup#5 @videlec
It is a solution but not great. Most users won't use version `n+1` of sage before one or two years (eg the time that the stable release make its way to Debian is even more than that). Which means that in the mean time they can not use your package.
