### 2022-03-22 fredstro/hilbertmodgroup#5 @fredstro
I understand that but I don't know of any perfect solution.

The most straightforward solution would be to go against PEP517 and delete pyproject.toml so pip uses the legacy build backend without build isolation.  This would probably work for at least a few years until the pip developers deprecate it.

Another more complicated option might be to setup some kind of buildbot to make wheels for as many platforms as possible. Since pip prefers installs from existing wheels before source distributions that should make it possible for most users to just do `sage -pip install` without the `--no-build-isolation` flag.
