### 2022-03-22 #3996 @fredstro
I have  addressed all issues except https://github.com/fredstro/hilbertmodgroup/issues/5
for which I don't really know what the best solution is.
It is a really complex question about Python packaging and pip build isolation.  I don't know if @videlec  agrees but for the purpose of the publication in JOSS I think that any of my proposed solutions should be sufficient and the search for an optimal solution if there even is one could be left for later.
