+++ paper.bib
@@ -1,7 +1,7 @@
 @manual{sage,
 	author = "{The Sage Developers}",
-	title = "{S}age {M}athematics {S}oftware ({V}ersion 9.4)",
-	year = 2021,
+	title = "{S}age {M}athematics {S}oftware ({V}ersion 9.5)",
+	year = 2022,
 	url         = "https://www.sagemath.org"
 }
 
@@ -26,14 +26,18 @@
   url="https://doi.org/10.1109/mcse.2010.118"
 }
 
-@misc{reductionalghilbert,
-      title={A Reduction Algorithm for {H}ilbert Modular Groups},
-      author={Strömberg, Fredrik},
-      year={2021},
-      eprint={2111.13274},
-      archivePrefix={arXiv},
-      primaryClass={math.NT}
+@article{reductionalghilbert,
+    title = {A reduction algorithm for {H}ilbert modular groups},
+    journal = {Journal of Number Theory},
+    year = {2022},
+    issn = {0022-314X},
+    doi = {10.1016/j.jnt.2022.02.011},
+    url = {https://www.sciencedirect.com/science/article/pii/S0022314X22000567},
+    author = {Strömberg, Fredrik},
+    keywords = {Hilbert modular groups, Fundamental domains, Reduction theory},
+    abstract = {The aim of this paper is to present an explicit reduction algorithm for Hilbert modular groups over arbitrary totally real number fields. An implementation of the algorithm is available to download from [20]. The exposition is self-contained and sufficient details are given for the reader to understand how it works and implement their own version if desired.}
 }
+
 @article{MR4091535,
   author =        "Quinn, Joseph and Verjovsky, Alberto",
   journal =       "Geometriae Dedicata",
