### 2022-03-31 fredstro/hilbertmodgroup#9 @videlec
private/public
In your `examples/` folder you do call functions starting with underscore
- `_bound_for_sigma_norm`
- `_candidate_integers_sigma`
- `_bound_for_rho_embeddings`
If these are supposed to be called by users why making them start with an underscore?

Part of https://github.com/openjournals/joss-reviews/issues/3996
