### 2022-03-31 #3996 @fredstro
It is trivial to write down a set of generators but it is very hard to find a set of "side-pairing" generators and get out a presentation from the geometry (which is the most conceptual way of doing it over Q). The code can be used to do this numerically but it is not easy as the "bottom" of the fundamental domain is very complicated in general.
I am working on another algorithm (which will use parts of this one) to find a reduced set of elliptic fixed-points and associated elements, which will result in a presentation.

For the moment I tried to make the Hilbert modular group classes having the same structure as the subgroups of the modular group and I didn't want to clutter it too much, which is why I introduced the utility class "HilbertPullback" which contains all functionality regarding the reduction. So at the moment it would be:
sage: p = a random point in (H^2)^degree
sage: q = (some matrix in SL2OK) * p
sage: R = HilbertPullback(SL2OK)
sage: R.reduce(p) == R.reduce(q)
True

I was thinking of having a "reduce" function also in the HilbertModularGroup but then I would get a circular dependency so I am not sure what the best way is to do that.
