### 2022-03-31 #3996 @videlec
Using `HilbertPullback` is perfectly fine (this reduction is not attached to the group anyway, you did a choice of fundamental domain). Could you had this `P1.reduce` early in the `ExamplesK1.ipynb`?
