### 2022-03-31 fredstro/hilbertmodgroup#9 @fredstro
They are not intended to be used by end users but I include them in the "examples" files because they demonstrate the different parts that goes into the algorithm.
There are two separate intentions behind the examples, the first is for end users to see how to use the algorithms and the second is to demonstrate the bounds and the inner workings to make it easier to verify that the algorithm is working correctly (and as described in the paper).
