### 2022-03-31 fredstro/hilbertmodgroup#9 @videlec
Maybe you could add more comments in the notebook in order to distinguish what the user is supposed to use (ie the public API that we should use when writing code that depends on `hilbertmodgroup`) and what the curious researcher might find interesting about the implementation details.
