### 2022-03-31 fredstro/hilbertmodgroup#9 @fredstro
I think that's a good idea. What do you think about the following at the top. This is really all a "regular" end user is expected to use.
