### 2022-03-31 b639b3b
Add ngens (number of generators) and random element functions to the Hilbert modular group class
