### 2022-04-07 #3996 @videlec
@fredstro Concerning "Community guidelines" it would be good to mention in the README
1) how people could contribute to the software? (I guess opening issues and opening merge requests?)
2) Report issues or problems with the software? (I guess opening issues)
3) Seek support?
