### 2022-04-07 #3996 @videlec
@danielskatz The above comment together with https://github.com/fredstro/hilbertmodgroup/issues/9 are my last comments for the review. The hilbertmodgroup library is in very good shape.
