### 2022-04-07 fredstro/hilbertmodgroup#8 @fredstro
I think this is fine now.
