### 2022-04-07 fredstro/hilbertmodgroup#7 @fredstro
This is done.
