### 2022-04-07 fredstro/hilbertmodgroup#6 @videlec
You still have an implementation of `__call__`. Is that intentional?
