### 2022-04-07 #3996 @fredstro
I have addressed issue #9 and I have also updated the README file with community guidelines (all in "develop")
