### 2022-04-07 #3996 @videlec
What is the point of having a `develop` and a `main` branch (the latter being the default)? The tags contains the relevant pointers to the commit corresponding to specific versions.
