### 2022-04-07 fredstro/hilbertmodgroup#6 @videlec
I was looking at the `main` branch. It is very confusing that the current branch is develop and not main. If you keep that setup write it down in the README.
