### 2022-04-07 fredstro/hilbertmodgroup#6 @fredstro
My current workflow is to use the develop branch for development and bugfixes and once they are verified as "correct" I will merge into main.
