### 2022-04-07 fredstro/hilbertmodgroup#6 @fredstro
Once / if there are more develpers involved I will use a more refined git-flow -based approach with feature branches.
