### 2022-04-07 fredstro/hilbertmodgroup#6 @fredstro
I know that releases re tagged but in my (perhaps wrong) opinion the "main" branch should always be as functionally correct as possible while the develop or feature branches can contain flaws to share between developers for comments and improvements.
