### 2022-04-07 fredstro/hilbertmodgroup#6 @videlec
This is non-sandard in github development (but not a problem). The more standard workflow is
- write code in a temporary branch and open merge request
- merge in main once the CI passed

You should describe your workflow in the README and be more specific on what users are expected to do with their merge requests (eg should we base our code on `main` or `develop`?)
