### 2022-04-07 #3996 @fredstro
The point is to have a `develop` branch for development and bugfixes and once the development is verified as "correct" it will be merged into main. I am only using a single "develop" branch since I am the only developer but if more people gets involved I will use git-flow with feature branches.

Even with tags and pointers to specific versions there will be users that just clone the repo and try to instal the latest main branch so it is good to keep that as clean and working as possible in each commit.
I am probably treating "main" more as a "release" branch...

If you think this is a wrong workflow and have a better suggestion just let me know.
