### 2022-04-07 #3996 @videlec
It is not wrong but it ought to be documented.
