### 2022-04-07 fredstro/hilbertmodgroup#6 @fredstro
I see.
In my experience (as a consultant working on a few slightly larger but not massive commercial projects) the workflow that seems to be common in industry is to use short lived feature branches, based of a develop branch and once fnished they are merged back into the development branch. (This is the default `git-flow` configuration, only hotfixes are supposed to go directly into `main`)

The reason for this is that the development branch can contain many different new features that also have to be tested to work together and might not be developed in sequence.
Once the current develop branch (contaning a number of  new features) is confirmed as working by QA (or a manager) it can be merged into main as part of the next release.

I will update the README.
