### 2022-04-07 #3996 @fredstro
I have added documentation about use of branches. Let me know if it is not clear enough.
