### 2022-04-07 #3996 @danielskatz
@fredstro - I'm suggesting some small changes in https://github.com/fredstro/hilbertmodgroup/pull/10. In addition, I think that perhaps you should formally cite Guass' work in the third paragraph, rather than just mentioning it inline.
