### 2022-04-07 #3996 @fredstro
Ok. I have merged your suggestions and added a  formal reference to (the English translation) of Gauss' work.
