+++ paper.bib
@@ -63,3 +63,15 @@
   url =           "https://doi.org/10.1112/S1461157015000121",
   note =   "code is available from: https://github.com/mstreng/reduce"
 }
+
+@book {MR0197380,
+    AUTHOR = {Gauss, Carl Friedrich},
+     TITLE = {Disquisitiones arithmeticae},
+      NOTE = {Translated into English by Arthur A. Clarke, S. J},
+ PUBLISHER = {Yale University Press, New Haven, Conn.-London},
+      YEAR = {1966},
+     PAGES = {xx+472},
+   MRCLASS = {10.00 (01.60)},
+  MRNUMBER = {0197380},
+MRREVIEWER = {W. J. LeVeque},
+}
