+++ CHANGELOG.md
@@ -1,15 +1,25 @@
 # Changelog
 
+## v0.1.0
+- First release version.
 
-## v.0.0.2
+## v0.0.3
+- Updated coercion and pickling for extension types.
+- Allow points on the boundary in the UpperHalfPlaneProduct class
+- Added random_element function to the Hilbert modular group class
+- Updated example Jupyter notebooks.
+- Fixes suggested by pyflakes.
+- Updated README.md
+
+## v0.0.2
 - Added Changelog (this file)
 - Improved handling of arguments in make and Docker.
 - Updated README.md with examples
 - 
-## v.0.0.1
+## v0.0.1
 - Updated docker commands
 - Added gitHub actions 
 - Added tox environments
 
-## v.0.0.0
+## v0.0.0
 - Initial release
