### 2022-04-07 #3996 @fredstro
I have  created at tagged 0.1.0 release and uploaded it to PyPi as well as archived it in Zenodo. In Zenodo I uploaded both the source distribution and a zip file of the complete git repository at this tag from GitHub.
I have checked that the metadata is correct (and I included my ORCID).

The DOI is: 10.5281/zenodo.6422510
