### 2022-04-07 #3996 @fredstro
Thanks a lot to @tbirkandan  and @videlec for a lot of helpful comments and to @danielskatz for being an extremely helpful editor.
Especially the comments from @videlec will help a lot with my future software design and journal submissions.
